requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.2
Brotli==1.1.0
celery==5.5.3
cffi==2.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect2==0.8.0
Django==5.2.5
django-crontab==0.7.1
django-weasyprint==2.4.0
django-widget-tweaks==1.5.0
fonttools==4.60.1
kombu==5.5.4
packaging==25.0
pillow==12.0.0
prompt_toolkit==3.0.52
pycparser==2.23
pydyf==0.11.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
